fix(auth): harden OAuth refresh per review (#5)
Addresses gemini-code-assist review on PR #5:
- oauth.go: refresh now uses an http.Client with a 15s timeout instead of
  the timeout-less http.DefaultClient. The refresh runs while OAuthFetcher
  holds its write lock, so a hung upstream would otherwise stall every
  concurrent GetCredentials indefinitely.
- oauth.go: bound the error-response body read with io.LimitReader(_, 2048)
  to avoid unbounded reads of an untrusted upstream response.
- fetcher.go: defensive nil checks on creds in both the on-demand and
  background refresh paths before dereferencing fields.
- jwt.go: jwtClaims.Exp is int64 (idiomatic for the epoch exp claim),
  dropping the float-to-int conversion.

Tests: go build/vet/test ./... all pass; gofmt clean.

+++ internal/auth/jwt.go
@@ -13,7 +13,7 @@ import (
 // trusting separately-stored fields) is how the upstream codex-rs CLI behaves
 // and keeps us from ever using a token whose stored metadata has drifted.
 type jwtClaims struct {
-	Exp    float64 `json:"exp"`
+	Exp    int64 `json:"exp"`
 	OpenAI struct {
 		ChatGPTAccountID string `json:"chatgpt_account_id"`
 	} `json:"https://api.openai.com/auth"`
@@ -47,7 +47,7 @@ func parseJWTClaims(token string) (jwtClaims, bool) {
 // never worse than before.
 func AccessTokenExpiresAtMs(accessToken string, fallbackMs int64) int64 {
 	if c, ok := parseJWTClaims(accessToken); ok && c.Exp > 0 {
-		return int64(c.Exp) * 1000
+		return c.Exp * 1000
 	}
 	return fallbackMs
 }
